+++ src/main/java/ippo/assignment2/models/Direction.java
@@ -1,6 +1,7 @@
 package ippo.assignment2.models;
 
 /**
+ *
  * @since 0.1.2
  */
 public enum Direction {