+++ src/test/java/ippo/assignment2/services/PlayersTest.java
@@ -1,6 +1,6 @@
 package ippo.assignment2.services;
 
-import ippo.assignment2.collections.Items;
+import ippo.assignment2.collections.ItemsCollection;
 import ippo.assignment2.models.Direction;
 import ippo.assignment2.models.Player;
 import ippo.assignment2.models.Room;
@@ -9,10 +9,10 @@ import java.io.IOException;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class PlayersTest {
-    Players playersService;
+    PlayersService playersService;
 
     public PlayersTest() {
-        this.playersService = new Players();
+        this.playersService = new PlayersService();
     }
 
     @Test
@@ -20,7 +20,7 @@ public class PlayersTest {
         Player player = this.playersService.get(false);
         assertTrue(player instanceof Player);
         assertTrue(player.getDirection() instanceof Direction);
-        assertTrue(player.getItems() instanceof Items);
+        assertTrue(player.getItems() instanceof ItemsCollection);
         assertTrue(player.getRoom() instanceof Room);
     }
 }