+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 ## CHANGELOG
+
+
 ---
 
 ### Tag 0.3.2
 
 - Update the JavaDoc annotations for the controllers.
+- Update the JavaDoc annotations for the collections, factories and services.
 
 ---